refactor: options for checker
