+++ packages/module/src/modules/checker.ts
@@ -1,0 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { Nuxt } from '@nuxt/schema'
+import { ModuleOptions } from '../module'
+
+export function setupESLintChecker(options: ModuleOptions, nuxt: Nuxt) {
+  throw new Error('ESLint check is not implemented yet')
+}
